move verions atribute to config atribute store
