+++ src/client/src/store/config/combineReducers.ts
@@ -3,12 +3,14 @@ import detailsPage from "./pages/detailsPageReducer";
 import previewStatus from "./config/previewReducer";
 import validations from "./validations/setValidationsReducer";
 import isValidatingName from "./validations/validatingNameReducer";
+import versions from "./versions/reducer";
 
 const configReducer = combineReducers({
   detailsPage,
   previewStatus,
   validations,
-  isValidatingName
+  isValidatingName,
+  versions
 });
 
 export default configReducer;
