+++ src/client/src/store/config/configType.ts
@@ -1,4 +1,4 @@
-import { IVersionData } from "../versions/model";
+import { IVersionData } from "./versions/model";
 
 import { ISetVisitedPageAction, IResetVisitedPageAction, ISetPageAction, ISetDetails } from "./pages/model";
 import { IResetWizardAction } from "./config/model";
