+++ src/client/src/store/config/versions/model.ts
@@ -1,0 +1,7 @@
+import { IVersions } from "../../../types/version";
+import { WIZARD_INFO_TYPEKEYS } from "../../typeKeys";
+
+export interface IVersionData {
+  type: WIZARD_INFO_TYPEKEYS.GET_TEMPLATE_INFO;
+  payload: IVersions;
+}
