+++ src/client/src/store/templates/templatesType.ts
@@ -1,4 +1,4 @@
-import { IVersionData } from "../versions/model";
+import { IVersionData } from "../config/versions/model";
 
 import { ISetVisitedPageAction, IResetVisitedPageAction, ISetPageAction, ISetDetails } from "./pages/model";
 import { IResetWizardAction } from "./preview/model";
